Break out each page into a separate component
