+++ components/header.js
@@ -31,24 +31,16 @@ class Header extends Nanocomponent {
     const loginClass = cn({ [styles.active]: page === "login" });
     const signupClass = cn({ [styles.active]: page === "signup" });
 
-    switch (page) {
-      case "login":
-      case "signup":
-        return html`
-        <div class="${styles.header}">
-          ${settings.disable_signup
-            ? ""
-            : html`<button class="${styles.btn} ${styles.btnHeader} ${signupClass}"
-              onclick=${this.navigateSignupPage}>Sign Up</button>`}
-          <button class="${styles.btn} ${styles.btnHeader} ${loginClass}"
-          onclick=${this.navigateLoginPage}>Log In</button>
-        </div>
-      `;
-      default:
-        return html`
-          <div class="${styles.header}"><button class="${styles.btn} ${styles.btnHeader} ${styles.active}">${title}</button></div>
-        `;
-    }
+    return html`
+      <div class="${styles.header}">
+        ${settings.disable_signup
+          ? ""
+          : html`<button class="${styles.btn} ${styles.btnHeader} ${signupClass}"
+            onclick=${this.navigateSignupPage}>Sign Up</button>`}
+        <button class="${styles.btn} ${styles.btnHeader} ${loginClass}"
+        onclick=${this.navigateLoginPage}>Log In</button>
+      </div>
+    `;
   }
 
   update ({ page, settings }, emit) {
